add percentage to top catgories

+++ expense_tracker/lib/features/home/models/top_category.dart
@@ -13,5 +13,5 @@ class TopCategory {
   String name;
   int total;
   String? color;
-  int? percentage;
+  double? percentage;
 }
